IconLayout/Item: enabled drag & drop
Also added new useDragFile that allows adding drag & drop functionnality
to any component with just a simple function call :^)
