+++ src/components/layouts/TableLayout/components/Row/index.tsx
@@ -1,17 +1,12 @@
 import React from 'react'
-import { DragPreviewImage, useDrag } from 'react-dnd'
+import { useTranslation } from 'react-i18next'
 
 import type { DraggedObject, FileViewItem } from '$src/types'
 import { InlineEditEvent, ItemMouseEvent, makeEvent } from '$src/hooks/useLayout'
 import { Name } from '../Column/Name'
 import { Size } from '../Column/Size'
-import { createDragPreview } from '$src/components/layouts/TableLayout/utils'
-import { useTranslation } from 'react-i18next'
 import { useFileClick } from '$src/hooks/useFileClick'
-
-interface CollectedProps {
-    isDragging: boolean
-}
+import { useDragFile } from '$src/hooks/useDragFile'
 
 export interface RowProps {
     rowData: FileViewItem
@@ -35,13 +30,6 @@ export const Row = ({
     isDarkModeActive,
 }: RowProps): JSX.Element => {
     const dragProps = getDragProps(index)
-    const [{ isDragging }, drag, preview] = useDrag<DraggedObject, unknown, CollectedProps>({
-        type: 'file',
-        item: dragProps,
-        collect: (monitor) => ({
-            isDragging: !!monitor.isDragging(),
-        }),
-    })
     const { t } = useTranslation()
     const clickHandler = makeEvent(index, rowData, onRowClick)
     const doubleClickHandler = makeEvent(index, rowData, onRowDoubleClick)
@@ -51,16 +39,17 @@ export const Row = ({
         doubleClickHandler,
         rightClickHandler: contextMenuHandler,
     })
-    const dragPreview =
-        dragProps.dragFiles.length > 1
-            ? createDragPreview(t('DRAG.MULTIPLE', { count: dragProps.dragFiles.length }), isDarkModeActive)
-            : undefined
+
+    const { dragRef, dragPreview } = useDragFile({
+        isDarkModeActive,
+        dragProps,
+    })
 
     return (
         <>
-            {dragPreview && <DragPreviewImage connect={preview} src={dragPreview} />}
+            {dragPreview}
             <div
-                ref={drag}
+                ref={dragRef}
                 {...mouseProps}
                 style={{ width: '100%', height: '100%', alignItems: 'center', display: 'flex' }}
             >
